pyproject.toml
[project]
name = "naboo"
version = "0.1.0"
description = "AI-powered family robot — Strands agent for mBot2"
requires-python = ">=3.11"

dependencies = [
    "strands-agents>=0.1.0",
    "paho-mqtt>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "ollama>=0.4.0",
    "amqtt>=0.11.3",
]

[project.scripts]
naboo = "naboo.__main__:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["naboo"]
